pyproject.toml
[tool.poetry]
name = "unsplashbot"
version = "0.1.0"
description = ""
authors = ["Ian Andwati <andwatiian@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^2.3.3"
python-dotenv = "^1.0.0"
pytelegrambotapi = "^4.14.0"
requests = "^2.31.0"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.4.0"

[tool.isort]
profile = "black"

[tool.black]
line-length = 88
target-version = ["py311"]
trailing-comma = true


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
